Fixes #24: Resourceful won't run from Vagrant
Due to recent changes in one of the used libraries Resourceful now
needs to use PHP 5.4. This version was not available on the currently
used version of Ubuntu so I updated the version.

During the update the provision command failed on the libicu version
so I had to update that as well.